ensure that 'datastore_active' matches the actual datastore state, #171, QOL-9327
- On resource update, compare the flag to the actual state of the datastore, and update the flag if they don't match
